Reformat col and row to createBoard
Creation of row and column variables have been moved to the
function which creates the 2D array in the first place.
Being that the rows and columns are portions of the 2D
array, it makes sense to have them be created in the same
function that creates the entire 2D array.